requirements.txt
python-nmap
colorama
dnspython
requests